+++ pkg/apis/pingcap/v1alpha1/tidb_config_wraper.go
@@ -19,7 +19,7 @@ import (
 
 	"github.com/pingcap/errors"
 	"github.com/pingcap/tidb-operator/pkg/util/config"
-	"github.com/pingcap/tidb-operator/pkg/util/toml"
+	"github.com/pingcap/tidb-operator/pkg/apis/util/toml"
 	"k8s.io/apimachinery/pkg/util/json"
 )
