+++ pkg/apis/pingcap/v1alpha1/tiflash_config_wraper_test.go
@@ -21,7 +21,7 @@ import (
 	"github.com/google/go-cmp/cmp"
 	fuzz "github.com/google/gofuzz"
 	. "github.com/onsi/gomega"
-	"github.com/pingcap/tidb-operator/pkg/util/toml"
+	"github.com/pingcap/tidb-operator/pkg/apis/util/toml"
 	"sigs.k8s.io/yaml"
 )
